requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data processing
pandas==2.2.0
numpy==1.26.3

# Kafka
kafka-python==2.0.2

# Qdrant
qdrant-client==1.7.3

# Embeddings
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2
Pillow==10.2.0
openai-whisper==20231117

# SMTP
aiosmtplib==3.0.1
jinja2==3.1.3

# Google Gemini
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6
# Streamlit
streamlit==1.30.0

# Data visualization
plotly==5.18.0
pandas==2.2.0
numpy==1.26.3

# HTTP client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pillow==10.2.0


